+++ src/main/java/coupon/application/CouponService.java
@@ -1,0 +1,4 @@
+package coupon.application;
+
+public class CouponService {
+}
