+++ src/main/java/coupon/domain/CouponCategory.java
@@ -1,4 +1,4 @@
-package coupon.domain.coupon;
+package coupon.domain;
 
 public enum CouponCategory {
