+++ src/main/java/coupon/domain/CouponIssueDate.java
@@ -1,4 +1,4 @@
-package coupon.domain.coupon;
+package coupon.domain;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
