+++ src/test/java/coupon/domain/CouponIssueDateTest.java
@@ -1,4 +1,4 @@
-package coupon.domain.coupon;
+package coupon.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
