+++ src/test/java/coupon/domain/CouponMinOrderAmountTest.java
@@ -1,4 +1,4 @@
-package coupon.domain.coupon;
+package coupon.domain;
 
 import static org.assertj.core.api.Assertions.assertThatCode;
